refactor(core): extract refresh helpers
