+++ crates/blz-core/src/url_resolver.rs
@@ -1,29 +1,29 @@
-//! URL resolution utilities for smart llms.txt variant detection
+//! URL resolution utilities for smart llms.txt variant detection.
 //!
 //! Automatically prefers llms-full.txt over llms.txt when available, with
 //! fallback to exact URL if neither variant exists. Uses HEAD requests to
 //! check availability before fetching content.
 
-use anyhow::Result;
-use blz_core::{ContentType, Fetcher, SourceVariant};
 use tracing::{debug, warn};
 
-/// Result of URL resolution with variant and content info
+use crate::{ContentType, Error, Fetcher, Result, SourceVariant};
+
+/// Result of URL resolution with variant and content info.
 #[derive(Debug, Clone)]
 pub struct ResolvedUrl {
-    /// The final URL that succeeded
+    /// The final URL that succeeded.
     pub final_url: String,
-    /// Which variant was resolved
+    /// Which variant was resolved.
     pub variant: SourceVariant,
-    /// Content type based on line count
+    /// Content type based on line count.
     pub content_type: ContentType,
-    /// Number of lines in the content
+    /// Number of lines in the content.
     pub line_count: usize,
-    /// Whether to warn the user about this source
+    /// Whether to warn the user about this source.
     pub should_warn: bool,
 }
 
-/// Resolve the best URL for llms.txt documentation
+/// Resolve the best URL for llms.txt documentation.
 ///
 /// Tries URLs in order:
 /// 1. llms-full.txt variant (if base URL ends in .txt)
@@ -31,24 +31,7 @@ pub struct ResolvedUrl {
 /// 3. Exact URL as provided
 ///
 /// Uses HEAD requests to check availability before fetching content.
-///
-/// # Examples
-///
-/// ```ignore
-/// use crate::utils::url_resolver::resolve_best_url;
-/// use blz_core::Fetcher;
-///
-/// # async fn example() -> anyhow::Result<()> {
-/// let fetcher = Fetcher::new()?;
-/// let resolved = resolve_best_url(&fetcher, "https://react.dev/llms.txt").await?;
-///
-/// // Will try llms-full.txt first, fallback to llms.txt
-/// assert_eq!(resolved.variant, blz_core::SourceVariant::LlmsFull);
-/// # Ok(())
-/// # }
-/// ```
 pub async fn resolve_best_url(fetcher: &Fetcher, base_url: &str) -> Result<ResolvedUrl> {
-    // Try variants in order of preference
     let variants = [
         try_full_variant(base_url),
         Some(base_url.to_string()),
@@ -57,7 +40,6 @@ pub async fn resolve_best_url(fetcher: &Fetcher, base_url: &str) -> Result<Resol
 
     for (idx, maybe_url) in variants.iter().enumerate() {
         if let Some(url) = maybe_url {
-            // Check if this URL is available
             let should_fetch = match fetcher.head_metadata(url).await {
                 Ok(head_info) => {
                     let status = head_info.status;
@@ -89,7 +71,6 @@ pub async fn resolve_best_url(fetcher: &Fetcher, base_url: &str) -> Result<Resol
                 continue;
             }
 
-            // URL exists or HEAD is unsupported; fetch the content to analyze it
             let (content, _sha256) = match fetcher.fetch(url).await {
                 Ok(result) => result,
                 Err(err) => {
@@ -101,11 +82,9 @@ pub async fn resolve_best_url(fetcher: &Fetcher, base_url: &str) -> Result<Resol
             let line_count = content.lines().count();
             let (content_type, should_warn) = classify_content(line_count);
 
-            // Determine variant based on which URL succeeded
             let variant = match idx {
-                0 => SourceVariant::LlmsFull, // llms-full.txt variant
+                0 => SourceVariant::LlmsFull,
                 1 => {
-                    // Exact URL - determine if it's a known variant
                     if base_url.ends_with("llms-full.txt") {
                         SourceVariant::LlmsFull
                     } else if base_url.ends_with("llms.txt") {
@@ -114,7 +93,7 @@ pub async fn resolve_best_url(fetcher: &Fetcher, base_url: &str) -> Result<Resol
                         SourceVariant::Custom
                     }
                 },
-                2 => SourceVariant::Llms, // llms.txt fallback
+                2 => SourceVariant::Llms,
                 _ => SourceVariant::Custom,
             };
 
@@ -128,16 +107,13 @@ pub async fn resolve_best_url(fetcher: &Fetcher, base_url: &str) -> Result<Resol
         }
     }
 
-    // None of the variants worked
-    anyhow::bail!(
+    Err(Error::NotFound(format!(
         "Failed to resolve any llms.txt variant for '{base_url}'. \
          Tried: llms-full.txt, exact URL, and llms.txt fallback."
-    )
+    )))
 }
 
-/// Generate llms-full.txt variant URL
 fn try_full_variant(url: &str) -> Option<String> {
-    // If URL ends with "llms.txt" (but not "llms-full.txt"), try "llms-full.txt"
     if url.ends_with("llms.txt") && !url.ends_with("llms-full.txt") {
         Some(url.replace("llms.txt", "llms-full.txt"))
     } else {
@@ -145,9 +121,7 @@ fn try_full_variant(url: &str) -> Option<String> {
     }
 }
 
-/// Generate llms.txt variant URL
 fn try_base_variant(url: &str) -> Option<String> {
-    // If URL ends with "llms-full.txt", try "llms.txt"
     if url.ends_with("llms-full.txt") {
         Some(url.replace("llms-full.txt", "llms.txt"))
     } else {
@@ -155,12 +129,11 @@ fn try_base_variant(url: &str) -> Option<String> {
     }
 }
 
-/// Determine content type from line count
 const fn classify_content(line_count: usize) -> (ContentType, bool) {
     if line_count > 1000 {
         (ContentType::Full, false)
     } else if line_count < 100 {
-        (ContentType::Index, true) // Warn user
+        (ContentType::Index, true)
     } else {
         (ContentType::Mixed, false)
     }
@@ -172,66 +145,25 @@ mod tests {
 
     #[test]
     fn test_try_full_variant() {
-        // Should convert llms.txt to llms-full.txt
         assert_eq!(
             try_full_variant("https://example.com/llms.txt"),
             Some("https://example.com/llms-full.txt".to_string())
         );
 
-        // Should not modify llms-full.txt
         assert_eq!(try_full_variant("https://example.com/llms-full.txt"), None);
 
-        // Should not modify custom URLs
         assert_eq!(try_full_variant("https://example.com/docs.txt"), None);
     }
 
     #[test]
     fn test_try_base_variant() {
-        // Should convert llms-full.txt to llms.txt
         assert_eq!(
             try_base_variant("https://example.com/llms-full.txt"),
             Some("https://example.com/llms.txt".to_string())
         );
 
-        // Should not modify llms.txt
         assert_eq!(try_base_variant("https://example.com/llms.txt"), None);
 
-        // Should not modify custom URLs
         assert_eq!(try_base_variant("https://example.com/docs.txt"), None);
     }
-
-    #[test]
-    fn test_classify_content() {
-        // Full documentation
-        let (content_type, should_warn) = classify_content(1500);
-        assert_eq!(content_type, ContentType::Full);
-        assert!(!should_warn);
-
-        // Index file (should warn)
-        let (content_type, should_warn) = classify_content(50);
-        assert_eq!(content_type, ContentType::Index);
-        assert!(should_warn);
-
-        // Mixed content
-        let (content_type, should_warn) = classify_content(500);
-        assert_eq!(content_type, ContentType::Mixed);
-        assert!(!should_warn);
-
-        // Edge cases
-        let (content_type, _) = classify_content(100);
-        assert_eq!(content_type, ContentType::Mixed);
-
-        let (content_type, _) = classify_content(1000);
-        assert_eq!(content_type, ContentType::Mixed);
-
-        let (content_type, _) = classify_content(1001);
-        assert_eq!(content_type, ContentType::Full);
-
-        let (content_type, should_warn) = classify_content(99);
-        assert_eq!(content_type, ContentType::Index);
-        assert!(should_warn);
-    }
-
-    // Integration tests with mock server would go here
-    // Requires tokio and wiremock, similar to fetcher tests
 }
